+++ packages/sh-design/src/components/seamless-scroll/index.ts
@@ -1,0 +1,9 @@
+import { withInstall } from '../../utils/install'
+import SeamlessScroll from './src/seamless-scroll.vue'
+
+export const ShSeamlessScroll = withInstall(SeamlessScroll)
+export default ShSeamlessScroll
+
+export * from './src/seamless-scroll'
+
+export type SeamlessScrollInstance = InstanceType<typeof SeamlessScroll>